perf kbuild: add generated Kconfig build-test cases
Thanks to the python module kconfiglib, we were able to generate build
configurations. The set is, of course, not exhaustive but for each
option, all the possible states are tested. If, during development
step, a new symbol is added into the Kconfig file, the target
build-test will not fail (olddefconfig is called so that new
parameters are filled with their default value).

Signed-off-by: Alexis Berlemont <alexis.berlemont@gmail.com>